requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
frozenlist==1.3.3
grpcio==1.53.0
grpcio-tools==1.53.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
openai==0.27.5
Pillow==9.5.0
platformdirs==3.5.0
protobuf==4.21.12
pygame-ce==2.2.1
pylint==2.17.4
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
stability-sdk==0.7.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
